+++ tests/testthat/test-codebook.R
@@ -34,7 +34,7 @@ test_that("warnings", {
   # all valid properties
   vardesc <- list("description" = c(id = "Subject ID"), 
                   "privacy" = c(T, F), 
-                  "type" = c("factor", "float"),
+                  "type" = c("string", "float"),
                   "propertyID" = c(id = "ID7"), 
                   "minValue" = c(y = -100), 
                   "maxValue" = c(y = 100),
@@ -44,16 +44,16 @@ test_that("warnings", {
                   "naValues" = c(id = "NOPE"),
                   "alternateName" = c(id = "part_id"), 
                   "unitCode" = c(id = "huh?"))
-  expect_silent(cb <- codebook(data, "data", vardesc, as_json = FALSE))
+  expect_silent(cb <- codebook(data, "data", vardesc, return = "list"))
   
   expect_equal(cb$variableMeasured[[1]], 
                list(`@type` = "PropertyValue",
                      name = "id",
                      description ="Subject ID",
                      privacy =TRUE,
-                     type ="factor",
+                     type ="string",
                      propertyID = "ID7",
-                     levels = LETTERS,
+                     levels = as.list(LETTERS),
                      ordered = TRUE,
                      na = "NA",
                      naValues = "NOPE",
@@ -71,13 +71,13 @@ test_that("warnings", {
   
   # add an invalid property
   vardesc$invalid <- c(id = "STOP")
-  expect_warning(cb <- codebook(data, "data", vardesc, as_json = FALSE),
+  expect_warning(cb <- codebook(data, "data", vardesc, return = "list"),
                  "The following variable properties are not standard: invalid", fixed = TRUE)
 })
 
 test_that("no vardesc", {
   data <- sim_design(2, 2, seed = 8675309, plot = FALSE)
-  cb <- codebook(data, as_json = FALSE)
+  cb <- codebook(data, return = "list")
   
   vm <- cb$variableMeasured
   
@@ -94,7 +94,7 @@ test_that("no vardesc", {
                              name = "B",
                              description = "B",
                              levels = list(B1 = "B1", B2 = "B2"),
-                             type = "factor",
+                             type = "string",
                              ordered = FALSE))
   
   expect_equal(vm[[3]], list(`@type` = "PropertyValue",
@@ -117,11 +117,11 @@ test_that("named factor levels", {
                      id = list(id = "Subject ID"),
                      plot = FALSE)
   expect_message(cb <- codebook(data), "id set to type string")
-  expect_message(cb <- codebook(data), "pet set to type factor")
-  expect_message(cb <- codebook(data, as_json = FALSE), "y set to type float")
+  expect_message(cb <- codebook(data), "pet set to type string")
+  expect_message(cb <- codebook(data, return = "list"), "y set to type float")
   
   output <- capture_output(print(cb))
-  expect_equal(output, "Codebook for data (Psych-DS 0.1.0)\n* id (string): Subject ID\n* pet (factor)\n  * Levels\n    * cat: Has a cat\n    * dog: Has a dog\n  * Ordered: FALSE\n* y (float): Happiness Score")
+  expect_equal(output, "Codebook for data (Psych-DS 0.1.0)\n========================================\n\n===== Dataset Parameters =====\n* name: data\n* schemaVersion: Psych-DS 0.1.0\n===== Column Parameters =====\n* id (string): Subject ID\n* pet (string)\n  * Levels\n    * cat: Has a cat\n    * dog: Has a dog\n  * Ordered: FALSE\n* y (float): Happiness Score")
 })
 
 test_that("with vardesc", {
@@ -132,10 +132,10 @@ test_that("with vardesc", {
                                   A2 = "Condition 2"),
                   levels = list(B = c(B1 = "First level",
                                       B2 = "Second level")))
-  cb <- codebook(data, "My Data", vardesc, as_json = FALSE)
+  cb <- codebook(data, "My Data", vardesc, return = "list")
   
   output <- capture_output(print(cb))
-  expect_equal(output, "Codebook for My Data (Psych-DS 0.1.0)\n* id (string): Subject ID\n* B (factor): Between-subject factor\n  * Levels\n    * B1: First level\n    * B2: Second level\n  * Ordered: FALSE\n* A1 (float): Condition 1\n* A2 (float): Condition 2")
+  expect_equal(output, "Codebook for My Data (Psych-DS 0.1.0)\n========================================\n\n===== Dataset Parameters =====\n* name: My Data\n* schemaVersion: Psych-DS 0.1.0\n===== Column Parameters =====\n* id (string): Subject ID\n* B (string): Between-subject factor\n  * Levels\n    * B1: First level\n    * B2: Second level\n  * Ordered: FALSE\n* A1 (float): Condition 1\n* A2 (float): Condition 2")
   
   # unseen levels
   vardesc <- list(description = c(id = "Subject ID",
@@ -145,7 +145,7 @@ test_that("with vardesc", {
                   levels = list(B = c(B1 = "First level",
                                       B2 = "Second level",
                                       B3 = "Third level")))
-  cb <- codebook(data, "My Data", vardesc, as_json = FALSE)
+  cb <- codebook(data, "My Data", vardesc, return = "list")
   # should convert levels to a list
   expect_equal(cb$variableMeasured[[2]]$levels,
                list(B1 = "First level",
@@ -168,7 +168,67 @@ test_that("ignores extra vardesc", {
   # turn off messages to check there are no warnings
   faux_options(verbose = FALSE)
   expect_silent(
-    cb <- codebook(data, "My Data", vardesc, as_json = FALSE)
+    cb <- codebook(data, "My Data", vardesc, return = "list")
   )
   faux_options(verbose = TRUE)
+})
+
+test_that("conversion", {
+  data <- data.frame(
+    i = as.integer(1:10),
+    d = as.double(1:10),
+    s = LETTERS[1:10],
+    f = rnorm(10),
+    b = rep(c(T, F), 5),
+    l = rep(0:1, 5)
+  )
+  
+  # no explicit conversion
+  ndata <- codebook(data, return = "data")
+  expect_equal(typeof(ndata$i), "integer")
+  expect_equal(typeof(ndata$d), "integer")
+  expect_equal(typeof(ndata$s), "character")
+  expect_equal(typeof(ndata$f), "double")
+  expect_equal(typeof(ndata$b), "logical")
+  expect_equal(typeof(ndata$l), "integer")
+  
+  # convert all to string
+  vd <- list(type = rep("s", 6))
+  ndata <- codebook(data, vardesc = vd, return = "data")
+  expect_equal(typeof(ndata$i), "character")
+  expect_equal(typeof(ndata$d), "character")
+  expect_equal(typeof(ndata$s), "character")
+  expect_equal(typeof(ndata$f), "character")
+  expect_equal(typeof(ndata$b), "character")
+  expect_equal(typeof(ndata$l), "character")
+  
+  # convert all to int
+  vd <- list(type = rep("i", 6))
+  ndata <- codebook(data, vardesc = vd, return = "data")
+  expect_equal(typeof(ndata$i), "integer")
+  expect_equal(typeof(ndata$d), "integer")
+  expect_equal(typeof(ndata$s), "character")
+  expect_equal(typeof(ndata$f), "double")
+  expect_equal(typeof(ndata$b), "integer")
+  expect_equal(typeof(ndata$l), "integer")
+  
+  # convert all to float
+  vd <- list(type = rep("f", 6))
+  ndata <- codebook(data, vardesc = vd, return = "data")
+  expect_equal(typeof(ndata$i), "double")
+  expect_equal(typeof(ndata$d), "double")
+  expect_equal(typeof(ndata$s), "character")
+  expect_equal(typeof(ndata$f), "double")
+  expect_equal(typeof(ndata$b), "double")
+  expect_equal(typeof(ndata$l), "double")
+  
+  # convert all to bool
+  vd <- list(type = rep("b", 6))
+  ndata <- codebook(data, vardesc = vd, return = "data")
+  expect_equal(typeof(ndata$i), "integer")
+  expect_equal(typeof(ndata$d), "double")
+  expect_equal(typeof(ndata$s), "character")
+  expect_equal(typeof(ndata$f), "double")
+  expect_equal(typeof(ndata$b), "logical")
+  expect_equal(typeof(ndata$l), "logical")
 })